LayoutTests/inspector/elements/breadcrumb-updates.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function changeClass()
{
    document.getElementsByClassName("firstClass")[0].className = "anotherClass";
}

function deleteClass()
{
    document.getElementsByClassName("anotherClass")[0].className = "";
}

function test()
{
    WebInspector.showPanel("elements");
    InspectorTest.expandElementsTree(step0);

    function step0()
    {
        InspectorTest.selectNodeWithId("target", step1);
    }

    function step1()
    {
        dumpBreadcrumb("Original breadcrumb");
        InspectorTest.evaluateInPage("changeClass()", step2);
    }

    function step2()
    {
        dumpBreadcrumb("After class change");
        InspectorTest.evaluateInPage("deleteClass()", step3);
    }

    function step3()
    {
        dumpBreadcrumb("After class removal");
        InspectorTest.completeTest();
    }

    function dumpBreadcrumb(message)
    {
        InspectorTest.addResult(message + ":");
        var result = [];
        var crumbs = WebInspector.panel("elements").crumbsElement;
        var crumb = crumbs.firstChild;
        while (crumb) {
            result.unshift(crumb.textContent);
            crumb = crumb.nextSibling;
        }
        InspectorTest.addResult(result.join(" > "));
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that breadcrumbs are updated upon involved element's attribute changes in the Elements panel.
</p>

<div class="firstClass">
    <div id="target"></div>
</div>

</body>
</html>
